% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmapdb.r
\name{depmapdb_getgenedependency}
\alias{depmapdb_getgenedependency}
\title{depmapdb_getgenedependency function allows the user to query a certain geneset in specific diseases to extract the probability of the gene dependency in the specific cancer phenotype.
The input to this function is the cancer phenotype and the genesymbols to query.
The output of this function is a dataframe with the depmapid, the genesymbol and the probability.}
\usage{
depmapdb_getgenedependency(depmapids = c(), genesymbols = c(), o = list())
}
\description{
depmapdb_getgenedependency function allows the user to query a certain geneset in specific diseases to extract the probability of the gene dependency in the specific cancer phenotype.
The input to this function is the cancer phenotype and the genesymbols to query.
The output of this function is a dataframe with the depmapid, the genesymbol and the probability.
}
