% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.r
\name{enrich_cirfess_spc}
\alias{enrich_cirfess_spc}
\title{CIRFESS is a protein localization database that has a confidence score related to the protein being a surface membrane protein.
The input to the function is a vector or gene/protein names or an expression data-set
The output will include the confidence score associated to each gene in the expression data-set}
\usage{
enrich_cirfess_spc(d)
}
\description{
CIRFESS is a protein localization database that has a confidence score related to the protein being a surface membrane protein.
The input to the function is a vector or gene/protein names or an expression data-set
The output will include the confidence score associated to each gene in the expression data-set
}
