% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.r
\name{enrich_compartments_sp}
\alias{enrich_compartments_sp}
\title{Compartments is a protein localization database that has a confidence score related to the protein being a surface membrane protein.
The input to the function is a vector or gene/protein names or an expression data-set
The output will include the confidence score associated to each gene in the expression dataset.}
\usage{
enrich_compartments_sp(d)
}
\description{
Compartments is a protein localization database that has a confidence score related to the protein being a surface membrane protein.
The input to the function is a vector or gene/protein names or an expression data-set
The output will include the confidence score associated to each gene in the expression dataset.
}
