% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.r
\name{enrich_depmap}
\alias{enrich_depmap}
\title{The Depmap project completed a CRISPR and siRNA screen of many genes across available cancer cell lines. The database has the probability that a gene is associated to cancer cell line proliferation. Given there are multiple cancer cell lines per cancer phenotype, the max probability is reported back for each gene.
The input to this function is an expression data-set and the cancer phenotype.
The output of this function is a column added to the expression data-set labeled depmap which includes the probability of the gene being associated to the cancer phenotype specified.}
\usage{
enrich_depmap(d, depmapids, collate = "merge", mergefunc = max, colname = NULL)
}
\description{
The Depmap project completed a CRISPR and siRNA screen of many genes across available cancer cell lines. The database has the probability that a gene is associated to cancer cell line proliferation. Given there are multiple cancer cell lines per cancer phenotype, the max probability is reported back for each gene.
The input to this function is an expression data-set and the cancer phenotype.
The output of this function is a column added to the expression data-set labeled depmap which includes the probability of the gene being associated to the cancer phenotype specified.
}
