% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.r
\name{enrich_go}
\alias{enrich_go}
\title{Gene Ontology is a functional database that links genes to biological pathways. The user can query which genes in an expression database are associated to pathways of interest using enrich_go.
The input to this function is an expression data-set and pathways of interest in parameter goids.
The output of this function is an added column to the expression data-set labeled go which includes a binary scoring based on whether the gene in the expression data-set is found in the GO pathways queried.}
\usage{
enrich_go(d, goids, collate = "separate", colname = NULL)
}
\description{
Gene Ontology is a functional database that links genes to biological pathways. The user can query which genes in an expression database are associated to pathways of interest using enrich_go.
The input to this function is an expression data-set and pathways of interest in parameter goids.
The output of this function is an added column to the expression data-set labeled go which includes a binary scoring based on whether the gene in the expression data-set is found in the GO pathways queried.
}
