% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.r
\name{enrich_opentarget}
\alias{enrich_opentarget}
\title{The Pediatric molecular target list is a list of targets with developed reagents that are associated to pediatric cancers. Users have the ability to query their expression data-set agaist this list by running it through enrich_opentargets.
The input to this function is an expression data-set.
The output of this function is a column added to the expression data-set labeled opentarget which is a binary score depending on whether or not the gene/protein from the expression data-set is found in this database.}
\usage{
enrich_opentarget(d)
}
\description{
The Pediatric molecular target list is a list of targets with developed reagents that are associated to pediatric cancers. Users have the ability to query their expression data-set agaist this list by running it through enrich_opentargets.
The input to this function is an expression data-set.
The output of this function is a column added to the expression data-set labeled opentarget which is a binary score depending on whether or not the gene/protein from the expression data-set is found in this database.
}
