% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.r
\name{enrich_uniprot}
\alias{enrich_uniprot}
\title{Uniprot is a protein database that includes information about protein localization, functionality and structure. Using enrich_uniprot, the user can query the extra-cellular length of proteins in an expression data-set.
The input to this function is the expression data-set.
The output of this function is a column added to the expression data-set labeled uniprot which includes the length of the extracellular section of the protein based on uniprot.}
\usage{
enrich_uniprot(d)
}
\description{
Uniprot is a protein database that includes information about protein localization, functionality and structure. Using enrich_uniprot, the user can query the extra-cellular length of proteins in an expression data-set.
The input to this function is the expression data-set.
The output of this function is a column added to the expression data-set labeled uniprot which includes the length of the extracellular section of the protein based on uniprot.
}
