% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.r
\name{project_load}
\alias{project_load}
\title{project_load is used when the user is wanting to load the cancer expression datasets into R. The user is able to summarize and re-scale the expression data-sets based on their input prior to running it through the program for target scoring. Using this function you are able to load more than a single data-set to summarize or analyze.
The input to this function is a yaml file which contains the path of the expression dataset file and all the needed configurations for summarizing and re-scaling. The input could also be an R generated project structure object which contains the prior information}
\usage{
project_load(yamlfile, loaddatasets = T, dosummarize = T, dorescale = T)
}
\description{
dosummarize = T - can calculate any or all the following parameters
}
\details{
\enumerate{
\item numhighexpressed -  how many samples have average or above average expression of protein
\item numexpressed -  how many samples have any expression of protein
\item mean - calculate mean expression of every protein across all samples
\item median - calculate median expression of every protein across all samples
dorescale = T - can rescale numbers using any of the following methods:
\item percentile
\item log2
\item z-score
\item tpm
\item rangemap0100 - assign numbers in vector (v) to values from 0-100 - (v - minv)/(maxv-minv) * 100)
The summarized data will be in p\link{'data'}
}

If dosummarize = F - the data will not be summarized and you can find  the names of the datasets in p\link{'datasets'}
}
