% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.r
\name{projects_optimweightsandcurves}
\alias{projects_optimweightsandcurves}
\title{projects_optimweightandcurves is used to optimize the weights and curves inputted by the user by maximizing the MAP algorithm score.
The use can select what attribute weights and curves need optimizing which can be inputted by listing the column names in the Iweights or Icurves parameters
Several optimization methods can be used and they include: checkzeros|null|Nelder-Mead|SANN|BFGS|CG|L-BFGS-B. One or several optimization methods can be used in a single run.
The optimization will start at the original weights used. The user can hard-maintain the sign of the weights by setting enforceweightsigns=T (not recommended), or more gracefully favor the maintainance of the signs by setting softenforceweightsigns=T (this is default).
To log the optimized weights in an excel file, the user can feed that using the optimlogfile= parameter.}
\usage{
projects_optimweightsandcurves(ps, ...)
}
\description{
projects_optimweightandcurves is used to optimize the weights and curves inputted by the user by maximizing the MAP algorithm score.
The use can select what attribute weights and curves need optimizing which can be inputted by listing the column names in the Iweights or Icurves parameters
Several optimization methods can be used and they include: checkzeros|null|Nelder-Mead|SANN|BFGS|CG|L-BFGS-B. One or several optimization methods can be used in a single run.
The optimization will start at the original weights used. The user can hard-maintain the sign of the weights by setting enforceweightsigns=T (not recommended), or more gracefully favor the maintainance of the signs by setting softenforceweightsigns=T (this is default).
To log the optimized weights in an excel file, the user can feed that using the optimlogfile= parameter.
}
