% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theratardb.r
\name{theratardb_disease2genesymbols}
\alias{theratardb_disease2genesymbols}
\title{theratardb_disease2genesymbols is used to query targets associated to specified cancer phenotypes present in the database.
The input to the function is the cancer phenotype. The user can also specify the mode of action of drug of interest using the moa parameter and the status of drug development using the status parameter.
The output is a data-frame from the database with all available metadata assoicated to genes linked to the disease. If user only wants to retrieve gene symbols, the asframe parameter can be set to FALSE.}
\usage{
theratardb_disease2genesymbols(
  diseases,
  asframe = T,
  moa = NULL,
  status = NULL,
  minstatusscore = NULL,
  exact = NULL,
  where = ""
)
}
\description{
theratardb_disease2genesymbols is used to query targets associated to specified cancer phenotypes present in the database.
The input to the function is the cancer phenotype. The user can also specify the mode of action of drug of interest using the moa parameter and the status of drug development using the status parameter.
The output is a data-frame from the database with all available metadata assoicated to genes linked to the disease. If user only wants to retrieve gene symbols, the asframe parameter can be set to FALSE.
}
