% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theratardb.r
\name{theratardb_moas}
\alias{theratardb_moas}
\title{theratardb_moas is used to query the various of mode of actions (MOAs) of drugs in the database
No input is required for this function.
The output is the list of distinct MOAs in the database.}
\usage{
theratardb_moas()
}
\description{
theratardb_moas is used to query the various of mode of actions (MOAs) of drugs in the database
No input is required for this function.
The output is the list of distinct MOAs in the database.
}
