% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theratardb.r
\name{theratardb_searchdisease}
\alias{theratardb_searchdisease}
\title{theratardb_searchdisease is used to query what diseases exist in the database
The input to this function is a disease key-word
The output is the list of diseases found that match the key-word exactly or have the key-word within the phrase.}
\usage{
theratardb_searchdisease(queries, exact = FALSE, where = "")
}
\description{
theratardb_searchdisease is used to query what diseases exist in the database
The input to this function is a disease key-word
The output is the list of diseases found that match the key-word exactly or have the key-word within the phrase.
}
